Added the necessary files to make the website work
The user must RUN the "entertainment_center_github.py" file to generate the website. The master files currently work as they should. However, the text that shows as a movie synopsis inside of the trailer popup window is just static text. I would like for it to be dynamically populated with text from the "entertainment_center_github.py" file that contains data for each movie on the website.